+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - button to cleanup localStorage invalid entries
 - Link object, major rewrite ensues
+- LibraryItem params conf, reduces or eliminates builder function necessity
 
 ## pre 2014-08-06